+++ domain/src/main/java/breezyweather/domain/weather/model/Pollen.kt
@@ -23,19 +23,35 @@ import java.io.Serializable
  */
 class Pollen(
     val alder: Int? = null,
+    val ash: Int? = null,
     val birch: Int? = null,
+    val chestnut: Int? = null,
+    val cypress: Int? = null,
     val grass: Int? = null,
+    val hazel: Int? = null,
+    val hornbeam: Int? = null,
+    val linden: Int? = null,
     val mold: Int? = null, // Not a pollen, but probably relevant to put with them
     val mugwort: Int? = null,
+    val oak: Int? = null,
     val olive: Int? = null,
+    val plane: Int? = null,
+    val plantain: Int? = null,
+    val poplar: Int? = null,
     val ragweed: Int? = null,
+    val sorrel: Int? = null,
     val tree: Int? = null,
+    val urticaceae: Int? = null,
+    val willow: Int? = null
 ) : Serializable {
 
     val isMoldValid: Boolean
         get() = mold != null
 
     val isValid: Boolean
-        get() = tree != null || alder != null || birch != null || grass != null
-            || olive != null || ragweed != null || mugwort != null || mold != null
+        get() = alder != null || ash != null || birch != null || chestnut != null ||
+                cypress != null || grass != null || hazel != null || hornbeam != null ||
+                linden != null || mold != null || mugwort != null || oak != null || olive != null ||
+                plane != null || plantain != null || poplar != null || ragweed != null ||
+                sorrel != null || tree != null || urticaceae != null || willow != null
 }